fix: Fixed threeOfAKind
Changed criteria when asking if hand is three of a kind, so now returns if it's three of a kind or full. It was comparing if the hand had three of a kind and later if it had a pair. Due a three of a kind has indeed three cards equals, it has two either. The bug was the criteria that assumed a hand having a three of a kind and a pair (even the pair was contained on the three of a kind) was a full.

Closes #4